Refactor tacotron2 tests and inherite style
